Makes pulp_file a pytest plugin
This way it can provide pulp_file fixtures to pulpcore anytime they are
needed.

[noissue]
